+++ src/main/managers/BaseSubTaskManager.ts
