+++ src/main/managers/PriorityTaskQueue.ts
