+++ src/main/managers/RecordingSubTaskManager.ts
