+++ src/main/types/recording.ts
